+++ packages/components/src/Calendar/Calendar.stories.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/rules-of-hooks */
 import { CalendarDate } from '@internationalized/date';
+import { useState } from '@storybook/preview-api';
 import type { Meta, StoryObj } from '@storybook/react';
-import { useState } from 'react';
 import { DateValue } from 'react-aria-components';
 import { Calendar } from './Calendar';