+++ themes/theme-core/src/components/index.ts
@@ -36,3 +36,5 @@ export * from './TextArea.styles';
 export * from './Tooltip.styles';
 export * from './Underlay.styles';
 export * from './Popover.styles';
+export * from './XLoader.styles';
+export * from './Pagination.styles';